Improve deletion log

+++ syndicate/connection/sns_connection.py
@@ -149,7 +149,9 @@ class SNSConnection(object):
 
         :type topic_arn: str
         """
-        self.client.delete_topic(TopicArn=topic_arn)
+        # make get api call first, because the delete function is idempotent
+        if self.client.get_topic_attributes(topic_arn):
+            self.client.delete_topic(TopicArn=topic_arn)
 
     def remove_topic_by_name(self, topic_name):
         """ Remove topic by arn.
